# internal/handler/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SafeBites API",
    "description": "Backend API for SafeBites – scan food product labels, analyse ingredients, manage user preferences, scan history, and favourites.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase / Google ID token passed as Bearer in Authorization header."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "user not found" }
        }
      },
...
      "IngredientScore": {
        "type": "object",
        "description": "Safety score and reasoning for a single ingredient.",
        "properties": {
          "ingredient_name": { "type": "string", "example": "Enriched Flour" },
          "safety_score":    { "type": "string", "example": "6", "description": "Score 1–10 as a string; the AI may return a number which is coerced to string." },
          "reasoning":       { "type": "string", "example": "Contains refined carbohydrates with limited nutritional value." }
        }
      },
      "ScorerResult": {
        "type": "object",
        "description": "Aggregate safety score for a set of ingredients.",
        "properties": {
          "ingredient_scores": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/IngredientScore" }
          },
          "overall_score": { "type": "number", "format": "double", "example": 6.2, "description": "Weighted average safety score (0–10)." }
        }
      },
      "AnalyzeResponse": {
        "type": "object",
        "description": "Result of scanning and scoring the original product. Does not include alternative recommendations.",
        "properties": {
          "status":               { "type": "string", "example": "success" },
          "product_name":         { "type": "string", "example": "Ritz Crackers" },
          "ingredient_breakdown": { "$ref": "#/components/schemas/ScorerResult" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/analyze": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Analyse a product image",
        "description": "Upload a food product label image. The AI pipeline extracts the product name, searches for ingredients, and scores each ingredient for safety. Returns `ingredient_breakdown` for the original scanned product only — alternative recommendations are fetched separately via the Recommendations endpoint.",
        "operationId": "analyzeImage",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Product image file (max 10 MB). Supported types: JPEG, PNG, WEBP."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ingredient breakdown for the scanned product.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnalyzeResponse" }
              }
            }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "413": { "description": "Image too large" },
          "500": { "description": "Internal error" }
        }
      }
    },
...
  }
}